# postgres_api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/moneothingrawdatas/insert": {
            "post": {
                "description": "Insert data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "moneothingrawdatas"
                ],
                "summary": "Insert data",
                "parameters": [
                    {
                        "description": "Insert data",
                        "name": "valueseinsertrelationdtoarchdto",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.insertrelationdto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "integer",
                                "format": "int64"
                            }
                        }
                    }
                }
            }
        },
        "/moneothingrawdatas/timerange": {
            "post": {
                "description": "get moneothing by value",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "moneothingrawdatas"
                ],
                "summary": "get moneothing with rawdata",
                "parameters": [
                    {
                        "description": "get moneothing with rawdata by timerange",
                        "name": "timestamprangesearchdto",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.timestamprangesearchdto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.moneothingwithvalue"
                            }
                        }
                    }
                }
            }
        },
        "/moneothingrawdatas/timestamp": {
            "post": {
                "description": "get moneothing by timestamp",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "moneothingrawdatas"
                ],
                "summary": "get moneothing",
                "parameters": [
                    {
                        "description": "get moneothing by timestamp",
                        "name": "timestampsearchdto",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.timestampsearchdto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.moneothingwithvalue"
                            }
                        }
                    }
                }
            }
        },
        "/moneothingrawdatas/value": {
            "post": {
                "description": "get moneothing by value",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "moneothingrawdatas"
                ],
                "summary": "get moneothing",
                "parameters": [
                    {
                        "description": "get moneothing by value",
                        "name": "valuesearchdto",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.valuesearchdto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.moneothingwithvalue"
                            }
                        }
                    }
                }
            }
        },
        "/moneothings": {
            "post": {
                "description": "get moneothing",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "moneothings"
                ],
                "summary": "get moneothing",
                "parameters": [
                    {
                        "description": "get a moneothing",
                        "name": "moneothingsearchdto",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.moneothingsearchdto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.moneothing"
                        }
                    }
                }
            }
        },
...
        "/moneothingwithrawdatas/timerange": {
            "post": {
                "description": "get values for moneothing",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "moneothingwitrawdatas"
                ],
                "summary": "get values for moneothing in given timerange",
                "parameters": [
                    {
                        "description": "Rawdatas for timerange",
                        "name": "moneothingrawdatatimerangedto",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.moneothingrawdatatimerangedto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.moneothingwithvaluesviewmodel"
                        }
                    }
                }
            }
        },
        "/moneothingwithrawdatas/timestamp": {
            "post": {
                "description": "get values for moneothing",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "moneothingwitrawdatas"
                ],
                "summary": "get values for moneothing at timestamp",
                "parameters": [
                    {
                        "description": "Rawdata next to timerange",
                        "name": "moneothingrawdatatimestampdto",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.moneothingrawdatatimestampdto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.moneothingwithvaluesviewmodel"
                        }
                    }
                }
            }
        },
        "/rawdatas": {
            "post": {
                "description": "get rawdatas by value",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rawdatas"
                ],
                "summary": "Show rawdatas",
                "parameters": [
                    {
                        "description": "Get rawdatas",
                        "name": "valuesearchdto",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.valuesearchdto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.rawdataviewmodel"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "main.insertrelationdto": {
            "type": "object",
            "properties": {
                "bulkinsert": {
                    "type": "boolean"
                },
                "numberofdatasets": {
                    "type": "integer"
                },
                "thingid": {
                    "type": "string"
                },
                "time": {
                    "type": "string"
                },
                "uniqueidentifier": {
                    "type": "string"
                },
                "values": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "main.moneothing": {
            "type": "object",
            "properties": {
                "Data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/main.moneothingrawdata"
                    }
                },
                "displayname": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "thingid": {
                    "type": "string"
                },
                "uniqueidentifier": {
                    "type": "string"
                }
            }
        },
        "main.moneothingrawdata": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "moneothing": {
                    "$ref": "#/definitions/main.moneothing"
                },
                "rawdata": {
                    "$ref": "#/definitions/main.rawdata"
                },
                "rawdataid": {
                    "type": "integer"
                },
                "thingid": {
                    "type": "integer"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
        "main.moneothingrawdatatimerangedto": {
            "type": "object",
            "properties": {
                "fromtime": {
                    "type": "string"
                },
                "pagenumber": {
                    "type": "integer"
                },
                "pagesize": {
                    "type": "integer"
                },
                "thingid": {
                    "type": "string"
                },
                "totime": {
                    "type": "string"
                },
                "uniqueidentifier": {
                    "type": "string"
                }
            }
        },
        "main.moneothingrawdatatimestampdto": {
            "type": "object",
            "properties": {
                "thingid": {
                    "type": "string"
                },
                "time": {
                    "type": "string"
                },
                "uniqueidentifier": {
                    "type": "string"
                }
            }
        },
        "main.moneothingsearchdto": {
            "type": "object",
            "properties": {
                "pagenumber": {
                    "type": "integer"
                },
                "pagesize": {
                    "type": "integer"
                },
                "thingid": {
                    "type": "string"
                },
                "uniqueidentifier": {
                    "type": "string"
                }
            }
        },
        "main.moneothingwithvalue": {
            "type": "object",
            "properties": {
                "displayname": {
                    "type": "string"
                },
                "thingid": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                },
                "uniqueidentifier": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
        "main.moneothingwithvaluesviewmodel": {
            "type": "object",
            "properties": {
                "displayname": {
                    "type": "string"
                },
                "rawdatas": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/main.valuewithtimestampviewmodel"
                    }
                },
                "thingid": {
                    "type": "string"
                },
                "uniqueidentifier": {
                    "type": "string"
                }
            }
        },
        "main.rawdata": {
            "type": "object",
            "properties": {
                "Data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/main.moneothingrawdata"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "value": {
                    "type": "string"
                }
            }
        },
        "main.rawdataviewmodel": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "value": {
                    "type": "string"
                }
            }
        },
        "main.timestamprangesearchdto": {
            "type": "object",
            "properties": {
                "from": {
                    "type": "string"
                },
                "pagenumber": {
                    "type": "integer"
                },
                "pagesize": {
                    "type": "integer"
                },
                "to": {
                    "type": "string"
                }
            }
        },
        "main.timestampsearchdto": {
            "type": "object",
            "properties": {
                "lower": {
                    "type": "boolean"
                },
                "pagenumber": {
                    "type": "integer"
                },
                "pagesize": {
                    "type": "integer"
                },
                "time": {
                    "type": "string"
                }
            }
        },
        "main.valuesearchdto": {
            "type": "object",
            "properties": {
                "pagenumber": {
                    "type": "integer"
                },
                "pagesize": {
                    "type": "integer"
                },
                "value": {
                    "type": "string"
                }
            }
        },
        "main.valuewithtimestampviewmodel": {
            "type": "object",
            "properties": {
                "timestamp": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        }
    }
}